fix(seo): getOpportunites 500 — cast $3::float (pagerank) dans la requête candidats
Erreur prod : 'invalid input syntax for type integer: "0.0216..."'. Le param
$3 (maxPagerank, float) était utilisé dans '$3 > 0' -> PostgreSQL (via node-pg)
inférait $3 en integer, et la valeur décimale du pagerank ne se parsait pas.

Fix : $3::float (et $2::int pour le plancher d'impressions) pour fixer les
types sans ambiguïté. Les facteurs du score restent en flottant côté JS
(imprF/posF/maillageDeficit), arrondi final uniquement.

Validé en local (PostgreSQL 16 + node-pg) : ancienne requête -> erreur integer
reproduite à l'identique ; requête corrigée -> OK, candidats corrects.

Push != déploiement.

+++ backend/controllers/seoController.js
@@ -250,12 +250,12 @@ const seoController = {
                 value_score::float AS value_score,
                 gsc_impressions, gsc_clicks, gsc_position::float AS gsc_position
          FROM seo_pages
-         WHERE site_id = $1 AND COALESCE(gsc_impressions, 0) >= $2
+         WHERE site_id = $1 AND COALESCE(gsc_impressions, 0) >= $2::int
            AND (
              (gsc_position IS NOT NULL AND gsc_position BETWEEN 11 AND 50)
              OR COALESCE(inlinks_count, 0) <= 2
              OR health IN ('orpheline', 'affamee')
-             OR (internal_pagerank IS NOT NULL AND $3 > 0 AND internal_pagerank / $3 < 0.6)
+             OR (internal_pagerank IS NOT NULL AND $3::float > 0 AND internal_pagerank / $3::float < 0.6)
            )`,
         [siteId, minImpr, maxPr]
       );
